backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic[email]==2.7.0
pydantic-settings==2.7.0
email-validator==2.1.0
python-dateutil==2.8.2

# Logging
loguru==0.7.2

# Input sanitization
bleach==6.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# AI/ML dependencies for resume matching
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.0
scikit-learn==1.3.2
spacy==3.7.2
PyMuPDF==1.23.8
numpy==1.24.3
pandas==2.1.3
nltk==3.8.1
regex==2023.10.3
